pyproject.toml
[project]
name = "ring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "slack-bolt>=1.26.0",
    "uvicorn>=0.38.0",
]
